fix: Event id is no longer tracked by an in memory incrementor. Event id's are now generated and stored uniquely on the database which allows for consistency between sessions. POST method call has been adjusted but still not functional.
