--- src/app/timetable/components/calender/event-utils.js
@@ -1,5 +1,0 @@
-let eventIdCounter = 0;
-
-export function createEventId() {
-  return String(eventIdCounter++);
-}
